Test multiple Ruby versions in CI by default
Add support for multiple Rubies to our gem template. Default
to 2.5, 2.6, 2.7, which are the current MRI Ruby versions
which work out of the box with panolint and rubocop without
errors.

While jruby also works, we've decided to leave it out of the
default set, as we don't currently use it at Panorama, and
don't want to assume we'd invest the time to support it in
every project by default. But it can be added to individual
projects of course.

Github Actions configuration based on:
https://github.com/ruby/setup-ruby/blob/master/README.md