Specify development dependencies
Gems differentiate between 'production' and 'development'
dependencies, with the impact that production dependencies
are transitively depended on by any app that depends on
the gem.

Here we move the Rake and RSpec dependencies to be specified
as development dependencies in the gemspec file.

We can't do that for a panolint becauyse it is a git
dependency, so we move it to a 'development' group
in the Gemfile.